acomodeLosArchivosDelProyecto
