+++ client/src/App.tsx
@@ -1,11 +1,11 @@
 import { Routes, Route } from "react-router-dom";
 import Home from "./home/Home";
 import SignUpForm from "./views/signUpForm/SignUpForm";
-import Products from "./components/Product";
-import Help from "./components/Help";
-import LandingPage from "./views/LandingPage";
+import CreaTuCuchillo from "./components/crea-tu-cuchillo/CreaTuCuchillo";
+import Help from "./components/help/Help";
+import LandingPage from "./views/landingPage/LandingPage";
 import Login from "./views/login/Login";
-import Footer from "./components/Footer";
+import Footer from "./components/Footer/Footer";
 import "./app.css";
 
 const App = () => {
@@ -16,7 +16,7 @@ const App = () => {
         <Route path="Login" element={<Login />} />
         <Route path="/home" element={<Home />} />
         <Route path="/signup" element={<SignUpForm />} />
-        <Route path="/products" element={<Products />} />
+        <Route path="/creatucuchillo" element={<CreaTuCuchillo />} />
         <Route path="/help" element={<Help />} />
       </Routes>
       <Footer />
